+++ src/components/tables/dfs/nba/PlayerGroupsNba.js
@@ -1,6 +1,5 @@
-
 import React, { useState } from 'react';
-import { Button, TextField, Select, MenuItem, InputLabel, FormControl } from '@mui/material';
+import { Button, TextField } from '@mui/material';
 import {
     Table,
     TableBody,
@@ -11,9 +10,11 @@ import {
     Paper,
 } from '@material-ui/core';
 
-function NbaTeamGameStacks(props) {
+function PlayerGroupsNba(props) {
     const [activeGroupId, setActiveGroupId] = useState(null);
-    const [selectedGame, setSelectedGame] = useState('');
+    const [searchTerm, setSearchTerm] = useState("");
+    const [groupSearchTerm, setGroupSearchTerm] = useState("");
+
     const generateUniqueId = () => {
         return new Date().getTime() + "-" + Math.floor(Math.random() * 1000);
     };
@@ -21,14 +22,12 @@ function NbaTeamGameStacks(props) {
     const handleCreateGroup = () => {
         props.setGroups(prev => [...prev, {
             id: generateUniqueId(),
-            name: "New Game Stack",
-            stackType: 'singleGameStack',
-            // numPlayers: 3,
-            forGame: '',
-            forPositions: [],
+            stackType: 'playerGroup',
+            name: "New Player Group",
+            players: [],
             minFromGroup: 1,
             maxFromGroup: 4,
-            maxExposure: 100
+            maxExposure: 50,
         }]);
     };
 
@@ -76,6 +75,12 @@ function NbaTeamGameStacks(props) {
         ]);
     };
 
+    const displayedPlayers = props.filteredPlayers.filter(player =>
+        player.Nickname.toLowerCase().includes(searchTerm.toLowerCase())
+    );
+
+
+
     return (
         <div style={{
             display: 'flex',
@@ -89,14 +94,14 @@ function NbaTeamGameStacks(props) {
                     backgroundColor: 'white',
                     marginTop: 16,
                     padding: 16,
-                    overflowY: 'auto',
+                    overflowY: 'hidden',
                     marginRight: 8,
                     // boxShadow: '0 2px 8px rgba(26, 24, 27, 0.06)',
 
                 }}>
                 <Button onClick={handleCreateGroup}
                     style={{ backgroundColor: '#00203d' }}
-                    variant="contained">Create Game Stack</Button>
+                    variant="contained">Create Player Group</Button>
                 <div style={{ height: 'calc(90% - 36px)', overflowY: 'auto' }}>
                     {props.groups.map((group) => (
                         <div
@@ -124,21 +129,20 @@ function NbaTeamGameStacks(props) {
             {activeGroupId && props.groups[getActiveGroupIndex()] && (
                 <div
                     className="playerGroupLeftSideWrapper"
+
                     style={{
                         width: '70%',
                         backgroundColor: 'white',
-                        overflowY: 'auto',
+                        overflowY: 'hidden',
                         marginTop: 16,
                         padding: 16,
-                    }}
-                >
-                    <h4 style={{ marginBottom: 16, marginTop: 24 }}>Edit Group Settings</h4>
+                        // boxShadow: '0 2px 8px rgba(26, 24, 27, 0.06)',
 
+                    }}>
+                    <h4 style={{ marginBottom: 16, marginTop: 24 }}>Edit Group Settings</h4>
                     <TextField
-                        style={{ margin: '24px 0px' }}
+                        style={{ margin: 4 }}
                         label="Group Name"
-                        fullWidth
-                        variant="outlined"
                         value={props.groups[getActiveGroupIndex()]?.name || ''}
                         onChange={e => {
                             const newName = e.target.value;
@@ -149,15 +153,10 @@ function NbaTeamGameStacks(props) {
                             ]);
                         }}
                     />
-
-
-
                     <TextField
-                        style={{ margin: '24px 0px' }}
+                        style={{ margin: 4 }}
                         type="number"
-                        label="Min Players From Group"
-                        fullWidth
-                        variant="outlined"
+                        label="Min From Group"
                         value={props.groups[getActiveGroupIndex()]?.minFromGroup || ''}
                         onChange={e => {
                             const newMin = e.target.value;
@@ -168,13 +167,10 @@ function NbaTeamGameStacks(props) {
                             ]);
                         }}
                     />
-
                     <TextField
-                        style={{ margin: '24px 0px' }}
+                        style={{ margin: 4 }}
                         type="number"
-                        label="Max number of Players"
-                        fullWidth
-                        variant="outlined"
+                        label="Max From Group"
                         value={props.groups[getActiveGroupIndex()]?.maxFromGroup || ''}
                         onChange={e => {
                             const newMax = e.target.value;
@@ -185,14 +181,10 @@ function NbaTeamGameStacks(props) {
                             ]);
                         }}
                     />
-
-
                     <TextField
-                        style={{ margin: '24px 0px' }}
+                        style={{ margin: 4 }}
                         type="number"
                         label="Max Exposure"
-                        fullWidth
-                        variant="outlined"
                         value={props.groups[getActiveGroupIndex()]?.maxExposure || ''}
                         onChange={e => {
                             const newMaxExposure = e.target.value;
@@ -204,65 +196,84 @@ function NbaTeamGameStacks(props) {
                         }}
                     />
 
-                    <div style={{ marginTop: 24, marginBottom: 8 }}>
-                        <h4>Select Positions for Group</h4>
-                        <FormControl fullWidth variant="outlined" style={{ margin: 4 }}>
-                            <InputLabel>Positions</InputLabel>
-                            <Select
-                                multiple
-                                value={props.groups[getActiveGroupIndex()]?.forPositions || []}
-                                onChange={e => {
-                                    const selectedOptions = e.target.value;
-                                    props.setGroups(prev => [
-                                        ...prev.slice(0, getActiveGroupIndex()),
-                                        { ...prev[getActiveGroupIndex()], forPositions: selectedOptions },
-                                        ...prev.slice(getActiveGroupIndex() + 1),
-                                    ]);
-                                }}
-                                label="Positions"
-                            >
-                                {props.positions.map(position => (
-                                    <MenuItem key={position} value={position}>
-                                        {position}
-                                    </MenuItem>
-                                ))}
-                            </Select>
-                        </FormControl>
-                    </div>
+                    <div style={{ display: 'flex', height: '500px', flexDirection: 'row', justifyContent: 'space-between' }}>
+                        <div style={{ width: '48%' }}>
+                            <h4 style={{ marginTop: 24, marginBottom: 8 }}>Add Players to Group</h4>
+                            <TextField
+                                style={{ marginBottom: 16 }}
+                                label="Search Players"
+                                variant="outlined"
+                                value={searchTerm}
+                                onChange={e => setSearchTerm(e.target.value)}
+                                fullWidth
+                            />
+                            <TableContainer component={Paper} style={{ width: '100%', height: '90%', overflowY: 'auto' }}>
+                                <Table>
+                                    <TableHead>
+                                        <TableRow>
+                                            <TableCell>Use</TableCell>
+                                            <TableCell>Name</TableCell>
+                                            <TableCell>FPPG</TableCell>
+                                        </TableRow>
+                                    </TableHead>
+                                    <TableBody>
+                                        {displayedPlayers.length === 0 ? (
+                                            <TableRow>
+                                                <TableCell colSpan={3} align="center">
+                                                    No players found
+                                                </TableCell>
+                                            </TableRow>
+                                        ) : (
+                                            displayedPlayers.map((player, index) => (
+                                                <TableRow key={index}>
+                                                    <TableCell>
+                                                        <input
+                                                            type="checkbox"
+                                                            checked={props.groups[getActiveGroupIndex()].players.includes(player.Nickname)}
+                                                            onChange={(e) => handleCheckboxChange(player.Nickname, e.target.checked)}
+                                                        />
+                                                    </TableCell>
+                                                    <TableCell>{player.Nickname}</TableCell>
+                                                    <TableCell>{player.FPPG}</TableCell>
+                                                </TableRow>
+                                            ))
+                                        )}
+                                    </TableBody>
+                                </Table>
+                            </TableContainer>
+                        </div>
 
-                    <div style={{ width: '100%', marginTop: 16 }}>
-                        <h4>Select Game for Group</h4>
-                        <TextField
-                            select
-                            fullWidth
-                            variant="outlined"
-                            value={props.groups[getActiveGroupIndex()]?.forGame || ''}
-                            onChange={e => {
-                                const newGame = e.target.value;
-                                props.setGroups(prev => [
-                                    ...prev.slice(0, getActiveGroupIndex()),
-                                    { ...prev[getActiveGroupIndex()], forGame: newGame },
-                                    ...prev.slice(getActiveGroupIndex() + 1),
-                                ]);
-                            }}
-                        >
-                            <MenuItem value="">
-                                <em>Select a game</em>
-                            </MenuItem>
-                            {props.gameMatchups.map((game, index) => (
-                                <MenuItem key={index} value={game.gameMatchup}>
-                                    {game.gameMatchup}
-                                </MenuItem>
-                            ))}
-                        </TextField>
+                        <div style={{ width: '48%' }}>
+                            <h4 style={{ marginTop: 24, marginBottom: 8 }}>Current Group Players</h4>
+                            <TableContainer component={Paper} style={{ height: '90%', overflowY: 'auto' }}>
+                                <Table>
+                                    <TableHead>
+                                        <TableRow>
+                                            <TableCell>Name</TableCell>
+                                            <TableCell>FPPG</TableCell>
+                                        </TableRow>
+                                    </TableHead>
+                                    <TableBody>
+                                        {props.groups[getActiveGroupIndex()].players.map((playerName, index) => {
+                                            const player = props.filteredPlayers.find(p => p.Nickname === playerName) || { Nickname: 'Unknown', FPPG: 'Unknown' };
+                                            return (
+                                                <TableRow key={index}>
+                                                    <TableCell>{player.Nickname}</TableCell>
+                                                    <TableCell>{player.FPPG}</TableCell>
+                                                </TableRow>
+                                            );
+                                        })}
+                                    </TableBody>
+                                </Table>
+                            </TableContainer>
+                        </div>
                     </div>
 
                 </div>
             )}
-
-
         </div>
     );
 }
 
-export default NbaTeamGameStacks;
+export default PlayerGroupsNba;
+
